requirements.txt
Django==2.2.4
jupyter==1.0.0
notebook==6.2.0
djangorestframework==3.12.2
markdown==3.3.3
django-filter==2.4.0
